refactor: todoList 스토리북 더미데이터 정리
