--- src/components/organisms/todo-list/dummy.ts
@@ -1,95 +1,0 @@
-import { TodoItem } from "@/types/todo";
-
-export const sampleTodos: TodoItem[] = [
-  {
-    id: 1,
-    title: "넥스트 공부하기",
-    done: false,
-    userId: 1,
-    teamId: "quesddo",
-    createdAt: "",
-    updatedAt: "",
-    noteId: 1,
-    linkUrl: "url",
-    fileUrl: "file",
-    goal: null,
-  },
-  {
-    id: 2,
-    title: "스토리북 설정하기",
-    done: true,
-    userId: 1,
-    teamId: "quesddo",
-    createdAt: "",
-    updatedAt: "",
-    noteId: 1,
-    linkUrl: "url",
-    fileUrl: null,
-    goal: null,
-  },
-  {
-    id: 3,
-    title: "운동하기",
-    done: true,
-    userId: 1,
-    teamId: "quesddo",
-    createdAt: "",
-    updatedAt: "",
-    noteId: 1,
-    linkUrl: null,
-    fileUrl: null,
-    goal: null,
-  },
-  {
-    id: 4,
-    title: "자바스크립트 기초 챕터2 보기",
-    done: true,
-    userId: 1,
-    teamId: "quesddo",
-    createdAt: "",
-    updatedAt: "",
-    noteId: null,
-    linkUrl: null,
-    fileUrl: "file",
-    goal: null,
-  },
-  {
-    id: 5,
-    title: "리액트 공식문서 읽기",
-    done: false,
-    userId: 1,
-    teamId: "quesddo",
-    createdAt: "",
-    updatedAt: "",
-    noteId: null,
-    linkUrl: "url",
-    fileUrl: null,
-    goal: null,
-  },
-  {
-    id: 6,
-    title: "빨래하고 건조기 돌리기",
-    done: true,
-    userId: 1,
-    teamId: "quesddo",
-    createdAt: "",
-    updatedAt: "",
-    noteId: null,
-    linkUrl: null,
-    fileUrl: null,
-    goal: null,
-  },
-  {
-    id: 7,
-    title: "리액트 쿼리 구조화 해보기",
-    done: false,
-    userId: 1,
-    teamId: "quesddo",
-    createdAt: "",
-    updatedAt: "",
-    noteId: null,
-    linkUrl: null,
-    fileUrl: null,
-    goal: null,
-  },
-];
