+++ src/mocks/todo/todoMockData.ts
@@ -1,4 +1,4 @@
-import { TeamIdTodosGet200Response } from "@/types/types";
+import { TeamIdTodosGet200Response, TodoResponseDto } from "@/types/types";
 
 export const todoListMockData = (): TeamIdTodosGet200Response => ({
   todos: [
@@ -32,3 +32,97 @@ export const todoListMockData = (): TeamIdTodosGet200Response => ({
   totalCount: 2,
   nextCursor: null,
 });
+
+export const sampleTodos: TodoResponseDto[] = [
+  {
+    id: 1,
+    title: "넥스트 공부하기",
+    done: false,
+    userId: 1,
+    teamId: "quesddo",
+    createdAt: "2024-02-26T12:00:00Z",
+    updatedAt: "2024-02-26T12:00:00Z",
+    noteId: 1,
+    linkUrl: "https://github.com/Quesddo/Client",
+    fileUrl: "https://github.com/Quesddo/Client",
+    goal: null,
+  },
+  {
+    id: 2,
+    title: "스토리북 설정하기",
+    done: true,
+    userId: 1,
+    teamId: "quesddo",
+    createdAt: "2024-02-26T12:00:00Z",
+    updatedAt: "2024-02-26T12:00:00Z",
+    noteId: 1,
+    linkUrl: "https://github.com/Quesddo/Client",
+    fileUrl: null,
+    goal: null,
+  },
+  {
+    id: 3,
+    title: "운동하기",
+    done: true,
+    userId: 1,
+    teamId: "quesddo",
+    createdAt: "2024-02-26T12:00:00Z",
+    updatedAt: "2024-02-26T12:00:00Z",
+    noteId: 1,
+    linkUrl: null,
+    fileUrl: "https://github.com/Quesddo/Client",
+    goal: null,
+  },
+  {
+    id: 4,
+    title: "자바스크립트 기초 챕터2 보기",
+    done: true,
+    userId: 1,
+    teamId: "quesddo",
+    createdAt: "2024-02-26T12:00:00Z",
+    updatedAt: "2024-02-26T12:00:00Z",
+    noteId: null,
+    linkUrl: null,
+    fileUrl: "https://github.com/Quesddo/Client",
+    goal: null,
+  },
+  {
+    id: 5,
+    title: "리액트 공식문서 읽기",
+    done: false,
+    userId: 1,
+    teamId: "quesddo",
+    createdAt: "2024-02-26T12:00:00Z",
+    updatedAt: "2024-02-26T12:00:00Z",
+    noteId: null,
+    linkUrl: "url",
+    fileUrl: null,
+    goal: null,
+  },
+  {
+    id: 6,
+    title: "빨래하고 건조기 돌리기",
+    done: true,
+    userId: 1,
+    teamId: "quesddo",
+    createdAt: "2024-02-26T12:00:00Z",
+    updatedAt: "2024-02-26T12:00:00Z",
+    noteId: null,
+    linkUrl: null,
+    fileUrl: null,
+    goal: null,
+  },
+  {
+    id: 7,
+    title: "리액트 쿼리 구조화 해보기",
+    done: false,
+    userId: 1,
+    teamId: "quesddo",
+    createdAt: "2024-02-26T12:00:00Z",
+    updatedAt: "2024-02-26T12:00:00Z",
+    noteId: null,
+    linkUrl: null,
+    fileUrl: null,
+    goal: null,
+  },
+];
